Crate freezer

On-disk copy-on-write datastructure tool

Structs

BlakeWrap
Wraps Blake2b in the CryptoHash trait

Freezer
Storage for values of type T, using hash-function H and storing data in backend B

VoidHash
Dummy hash function used in tests

Enums

Location
A location in the freezer, either in memory or on disk.

Traits

Backend
A backend to store and retrieve bytes from

CryptoHash
A wrapper trait for a cryptographic hash function

Freeze
Write Self as bytes or read Self from bytes

Sink
A sink implementation

Type Definitions

DefaultFreezer
A Freezer using Blake2b and a filesystem tree as the backend