Trait freezer::Sink

pub trait Sink<H> where H: CryptoHash, Self: Write {
    fn bypass_hashing<'a>(&'a mut self) -> Bypass<'a, H>;
    fn fin(self: Box<Self>) -> Result<H::Digest>;
    fn recurse(&self, _: &Location<H>) -> Result<H::Digest>;
}

A sink implementation

Required Methods

fn bypass_hashing<'a>(&'a mut self) -> Bypass<'a, H>

Write to the underlying stream without affecting the hash-state

fn fin(self: Box<Self>) -> Result<H::Digest>

Consume the sink, returning the hash of the written data

fn recurse(&self, _: &Location<H>) -> Result<H::Digest>

Recursively freeze a value from within a parent.